Move IsBlockMatrix from block_indices.h to constraint_matrix.h.
The latter is the only file where it is actually used, so we may as well have it there.
